% Generated by roxygen2 (4.0.1): do not edit by hand
\name{.attributeToDouble}
\alias{.attributeToDouble}
\title{Converts XML attributes.}
\usage{
.attributeToDouble(attributes, attributeName, required = FALSE)
}
\arguments{
\item{attributes}{\code{character}, XML attributes}

\item{attributeName}{\code{character}, name XML attributes of attribute to
convert}

\item{required}{\code{logical}, throw an error if an required attribute is
 missing.}
}
\value{
\code{double}
}
\description{
This function converts a XML attribute to \code{double}.
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
\seealso{
\code{\link[readMzXmlData]{.attributeToString}},
 \code{\link[readMzXmlData]{.attributeTimeToDouble}}
}
\keyword{internal}

