% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grep-functions.R
\name{.grepDouble}
\alias{.grepDouble}
\title{Pattern matching.}
\usage{
.grepDouble(x)
}
\arguments{
\item{x}{\code{character}, text}
}
\value{
\code{double}, matched value
}
\description{
This function returns \code{double} values found in a \code{character}.
}
\seealso{
\code{\link[readMzXmlData]{.grepNumber}}
 \code{\link[readMzXmlData]{.grepSubString}}
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
\keyword{internal}
