% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grep-functions.R
\name{.grepNumber}
\alias{.grepNumber}
\title{Pattern matching.}
\usage{
.grepNumber(x)
}
\arguments{
\item{x}{\code{character}, text}
}
\value{
\code{numeric}, matched value
}
\description{
This function returns \code{numeric} values found in a \code{character}.
}
\seealso{
\code{\link[readMzXmlData]{.grepDouble}}
 \code{\link[readMzXmlData]{.grepSubString}}
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
\keyword{internal}
