% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grep-functions.R
\name{.grepSubString}
\alias{.grepSubString}
\title{Pattern matching.}
\usage{
.grepSubString(pattern, x)
}
\arguments{
\item{pattern}{\code{character}, regexp pattern}

\item{x}{\code{character}, text}
}
\value{
\code{character}, matched substring
}
\description{
This function returns a substring selected by a regular expression.
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
\seealso{
\code{\link[readMzXmlData]{.grepNumber}}
}
\keyword{internal}

