\name{.mzXmlHandlers}
\alias{.mzXmlHandlers}
\alias{.mzXmlHandlersHash}
\title{Parse mzXML files.}
\usage{
  .mzXmlHandlers(fileName, verbose = FALSE)

  .mzXmlHandlersHash(fileName, verbose = FALSE)
}
\arguments{
  \item{fileName}{\code{character}, path to mzXML file}

  \item{verbose}{\code{logical}, verbose output?}

  \item{fileName}{\code{character}, path to mzXML file}

  \item{verbose}{\code{logical}, verbose output?}
}
\value{
  function closure

  function closure
}
\description{
  This function is defines handlers for XML SAX parser.
  Internal use only.

  This function is defines handlers for XML SAX parser.
  Internal use only.
}
\author{
  Sebastian Gibb \email{mail@sebastiangibb.de}

  Sebastian Gibb \email{mail@sebastiangibb.de}
}
\references{
  Definition of \code{mzXML} format:
  \url{http://tools.proteomecenter.org/mzXMLschema.php}

  Definition of \code{mzXML} format:
  \url{http://tools.proteomecenter.org/mzXMLschema.php}
}
\keyword{internal}

