% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMzXml.R
\name{.parseMzXml}
\alias{.parseMzXml}
\title{Parse mzXML files.}
\usage{
.parseMzXml(file, verbose = FALSE, ...)
}
\arguments{
\item{file}{\code{character}, path to mzXML file}

\item{verbose}{\code{logical}, verbose output?}
}
\value{
Returns a list with metadata and spectra.
}
\description{
This function parses mzXML files.
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
\keyword{internal}
