% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{readMzXmlData-defunct}
\alias{readMzXmlData-defunct}
\alias{mqReadMzXml}
\title{Removed functions in package \pkg{readMzXmlData}}
\description{
These functions are defunct and no longer available.
}
\details{
\describe{
 \item{mqReadMzXml:}{use
   \code{\link[MALDIquantForeign]{importMzXml}} instead.}
}
}
\keyword{internal}
