\name{readMzXmlData-internal}

\alias{.globalMetaData}
\alias{.scanMetaData}

\title{Internal readMzXmlData Functions}
\description{
Internal readMzXmlData functions.
}
\note{
These are not to be called by the user (or in some cases are just
waiting for proper documentation to be written).
}
\keyword{internal}
