% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base64-functions.R
\name{.base64decode}
\alias{.base64decode}
\title{Converts base64 character to double.}
\usage{
.base64decode(
  x,
  size,
  endian = .Platform$endian,
  compressionType = c("none", "gzip")
)
}
\arguments{
\item{x}{\code{character}, base64 encoded string.}

\item{size}{\code{integer}, number of bytes per element in the byte stream
(see \code{size} in \code{\link[base]{readBin}}).}

\item{endian}{\code{character}, the endian-ness
(see \code{endian} in \code{\link[base]{readBin}}).}

\item{compressionType}{\code{character}, type of compression to use for
decompression of \code{x} (see \code{type} in
\code{\link[base]{memDecompress}}.}
}
\value{
Vector of type \code{double}.
}
\description{
This function converts a base64 encoded \code{character} vector to a
\code{double} vector.
}
\seealso{
\code{\link[base64enc]{base64decode}} from \pkg{base64enc} package
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
\keyword{internal}
