% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMzXml.R
\name{.mzXmlHandlers}
\alias{.mzXmlHandlers}
\title{Parse mzXML files.}
\usage{
.mzXmlHandlers(fileName, verbose = FALSE)
}
\arguments{
\item{fileName}{\code{character}, path to mzXML file}

\item{verbose}{\code{logical}, verbose output?}
}
\value{
function closure
}
\description{
This function is defines handlers for XML SAX parser. Internal use only.
}
\references{
Definition of \code{mzXML} format:
\url{http://tools.proteomecenter.org/wiki/index.php?title=Formats:mzXML}
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
\keyword{internal}
