% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{readMzXmlData-package}
\alias{readMzXmlData-package}
\title{The readMzXmlData Package}
\description{
The package reads mass spectrometry data in mzXML format. \cr
}
\details{
\tabular{ll}{

 Package: \tab readMzXmlData\cr

 Type: \tab Package\cr

 Version: \tab 2.8.1\cr

 Date: \tab 2015-09-16\cr

 License: \tab GPL(version 3 or later)\cr

}

Main functions:

\code{\link{readMzXmlFile}}: Reads mass spectrometry data in mzXML format.

\code{\link{readMzXmlDir}}: Reads recursively mass spectrometry data in mzXML
format in a specific directory.

\code{\link{mqReadMzXml}}: Reads mass spectrometry data into MALDIquant.
}
\references{
See website: \url{https://strimmerlab.github.io/software/maldiquant/}
}
\seealso{
\code{\link{readMzXmlDir}}, \code{\link{readMzXmlFile}}
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
\keyword{IO}
