\docType{package}
\name{readMzXmlData-package}
\alias{readMzXmlData-package}
\title{The readMzXmlData Package}
\description{
  The package reads mass spectrometry data in mzXML format.
  \cr
}
\details{
  \tabular{ll}{

  Package: \tab readMzXmlData\cr

  Type: \tab Package\cr

  Version: \tab 2.6\cr

  Date: \tab 2013-04-03\cr

  License: \tab GPL(version 3 or later)\cr

  }

  Main functions:

  \code{\link{readMzXmlFile}}: Reads mass spectrometry data
  in mzXML format.

  \code{\link{readMzXmlDir}}: Reads recursively mass
  spectrometry data in mzXML format in a specific
  directory.

  \code{\link{mqReadMzXml}}: Reads mass spectrometry data
  into MALDIquant.
}
\author{
  Sebastian Gibb \email{mail@sebastiangibb.de}
}
\references{
  See website:
  \url{http://strimmerlab.org/software/maldiquant/}
}
\seealso{
  \code{\link{readMzXmlDir}}, \code{\link{readMzXmlFile}},
  \code{\link{mqReadMzXml}}
}
\keyword{IO}

